requirements.txt
aiohttp
aiohttp-cors
aioshutil
aiofiles
anitopy
apscheduler
dnspython
feedparser
psutil
pycountry
python-dotenv
tgcrypto
git+https://github.com/kaif-00z/html-telegraph-poster
lxml
lxml[html_clean]
motor==3.3.2
pymongo==4.6.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydrive2
pyrofork==2.3.45
boto3

# ✅ নিচের গুলো নতুন অ্যাড করা হলো:
requests
uvloop
beautifulsoup4
